Remove unnecessary `expect/actual` arrangement
